Create Subway Surfers
In this program, we define an array treasureWeights to store the weights of the five treasures. We then initialize a variable totalWeight to zero to store the total weight collected. Using a for loop, we iterate through each treasure, adding its weight to totalWeight. Finally, we print the total weight collected.

Note: This program assumes that Mr. John collects all five treasures.