square and rectangle
In this program, we define a Shape structure to represent the common attributes (coordinates) of both Square and Rectangle. Then, we define Square and Rectangle structures, each embedding a Shape structure and adding their specific attributes. We also define functions to calculate the area of each shape.

In the main function, we create objects of Square and Rectangle, initialize their attributes, and demonstrate the usage of the functions to calculate their areas.